requirements.txt
streamlit==1.32.0
beautifulsoup4==4.12.3
requests==2.31.0
nltk==3.8.1
newspaper3k==0.2.8
lxml[html_clean]>=4.9.3
lxml_html_clean
transformers==4.38.2
torch==2.2.1
gTTS==2.5.1 